twitter_producer/app/requirements.txt
kafka-python==2.0.2
tweepy==4.10.1